# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Akasi Onboarding API",
        "description": "API de gestion de l'onboarding pour les applications du groupe Akasi. Cette API permet de provisionner des tenants, de gérer des webhooks et de suivre l'état de déploiement.",
        "contact": {
            "email": "tech@akasi-group.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de Développement Local"
        }
    ],
    "paths": {
        "/api/v1/onboarding/start": {
            "post": {
                "tags": [
                    "Onboarding Stateless"
                ],
                "summary": "Démarrer un onboarding stateless",
                "description": "Crée un enregistrement d'onboarding dans la base centrale et génère un sous-domaine.",
                "operationId": "98832ff01ce791a7505c538049b16193",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "organization_name": {
                                        "type": "string",
                                        "example": "Clinique du Lac",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Onboarding créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "uuid": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "subdomain": {
                                            "type": "string",
                                            "example": "clinique-du-lac"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "organization_name": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "onboarding_status": {
                                            "type": "string",
                                            "example": "pending"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Master key invalide ou absente"
                    },
                    "500": {
                        "description": "Erreur interne lors du démarrage de l'onboarding"
                    }
                },
                "security": [
                    {
                        "MasterKey": []
                    }
                ]
            }
        },
        "/api/v1/onboarding/provision": {
            "post": {
                "tags": [
                    "Onboarding Stateless"
                ],
                "summary": "Provisionner l'infrastructure d'un onboarding",
                "description": "Configure DNS/SSL pour le sous-domaine et génère éventuellement une clé API.",
                "operationId": "0dee283b14762f8f5a843127a5553b35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "uuid"
                                ],
                                "properties": {
                                    "uuid": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "generate_api_key": {
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Onboarding provisionné",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "uuid": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "subdomain": {
                                            "type": "string",
                                            "example": "clinique-du-lac"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "organization_name": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "onboarding_status": {
                                            "type": "string",
                                            "example": "activated"
                                        },
                                        "api_key": {
                                            "type": "string",
                                            "example": "ak_abc123...",
                                            "nullable": true
                                        },
                                        "api_secret": {
                                            "type": "string",
                                            "example": "ak_abc123...",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Onboarding introuvable pour cette application"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "401": {
                        "description": "Master key invalide ou absente"
                    },
                    "500": {
                        "description": "Erreur interne lors du provisioning"
                    }
                },
                "security": [
                    {
                        "MasterKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "MasterKey": {
                "type": "apiKey",
                "description": "Master key de votre application cliente, obtenue via /api/v1/applications/register",
                "name": "X-Master-Key",
                "in": "header"
            }
        }
    },
...
}